phrase1: Import Asus-merlin 380.70

+++ drivers/mtd/bcm947xx/nand/bcm_nflash.c
@@ -1,7 +1,7 @@
 /*
  * Broadcom SiliconBackplane chipcommon serial flash interface
  *
- * Copyright (C) 2012, Broadcom Corporation. All Rights Reserved.
+ * Copyright (C) 2013, Broadcom Corporation. All Rights Reserved.
  * 
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -544,7 +544,7 @@ nflash_mtd_init(void)
 	/* Scan bad block */
 	NFLASH_LOCK(&nflash);
 	for (i = 0; i < info->numblocks; i++) {
-		if (hndnand_checkbadb(nflash.nfl, (i * info->blocksize)) != 0) {
+		if (hndnand_checkbadb(nflash.nfl, (i * (uint64)info->blocksize)) != 0) {
 			nflash.map[i] = 1;
 		}
 	}
